Commit de "/" inicial para que cargue los estilos correctamente para paginas como empresas/filtrosListaEmpresas

+++ src/main/java/ar/edu/utn/frba/dds/tp/dominio/Metodologia.java
@@ -58,7 +58,7 @@ public class Metodologia {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
